toolkit/components/url-classifier/tests/mochitest/test_classify_by_default.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1442496</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1442496">Mozilla Bug 1442496</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

// To add a request to test, add the request in trackerFrame.html
// and the id of query string "?id=xxx" here.
const trackersAll = [
  "img-src",
  "object-data",
  "script-src",
  "iframe-src",
  "link-rel-stylesheet",
  "link-rel-prefetch",
  "video-src",
  "track-src",
  "ping",
  "fetch",
  "xmlhttprequest",
  "send-beacon",
];
const trackerFrame = "http://itisatracker.org/tests/toolkit/components/url-classifier/tests/mochitest/trackerFrame.sjs";

// This function ask the server to set the cookie
function setup() {
  return fetch(trackerFrame + "?init=" + trackersAll.length, {
    credentials: "include",
  });
}

function loadTestWindow(withcookie) {
  return new Promise(resolve => {
    let win;
    let query = withcookie ? "with-cookie" : "without-cookie";
    fetch(trackerFrame + "?callback=" + query, {
      credentials: "include",
    }).then(r => {
      r.text().then((body) => {
        let trackers_found = body.split(",");
        for (let tracker of trackersAll) {
          let description = "Tracker request " + tracker + "received " +
            (withcookie ? "with" : "without") + " cookie";
          ok(trackers_found.includes(tracker), description);
        }
        win.close();
        resolve();
      });
    });

    win = window.open("trackerFrame.html");
  });
}

async function runTests() {
  await SpecialPowers.pushPrefEnv({set: [
    [ "urlclassifier.trackingAnnotationTable.testEntries", "itisatracker.org" ],
    [ "network.cookie.cookieBehavior", 4],
    [ "privacy.trackingprotection.enabled", false ],
    [ "privacy.trackingprotection.annotate_channels", false],
    [ "browser.send_pings", true],
  ]});

  await setup();

  await loadTestWindow(true);

  await SpecialPowers.pushPrefEnv({set: [
    [ "privacy.trackingprotection.annotate_channels", true],
  ]});

  await loadTestWindow(false);

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
runTests();

</script>

</pre>
</body>
</html>
